pre_screening_v2/pyproject.toml
[project]
name = "elevenlabs-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[deepgram,elevenlabs,google,openai,silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
